Move form containers to pages folder
