+++ src/components/Forms/NewEventForm.js
@@ -8,10 +8,10 @@ import {
     MuiPickersUtilsProvider,
     KeyboardDateTimePicker
 } from '@material-ui/pickers';
-import ThemeProvider from '../../ThemeProvider'
+import ThemeProvider from '../ThemeProvider'
 const slugify = require('slugify')
 
-export default function EditEventForm(props) {
+export default function NewEventForm(props) {
     const {
         values: { ename, slug, description, capacity, elocation, imageUrl, startDate, endDate },
         errors,
@@ -30,7 +30,6 @@ export default function EditEventForm(props) {
 
     const handleEventNameChange = (name, e) => {
         e.persist();
-        console.log(props.event)
         const newSlug = slugify(e.target.value, { lower: true });
         setFieldValue('slug', newSlug)
         handleChange(e);
@@ -47,7 +46,7 @@ export default function EditEventForm(props) {
 
     return (
         <ThemeProvider>
-            <Typography variant="h1">Edit Event</Typography>
+            <Typography variant="h1">Create a New Event</Typography>
             <form onSubmit={handleSubmit}>
                 <Paper>
                     <TextField
@@ -101,6 +100,7 @@ export default function EditEventForm(props) {
                         <KeyboardDateTimePicker
                             margin="normal"
                             label="Start Date"
+                            minDate={new Date()}
                             value={startDate}
                             onChange={handleStartDateChange}
                         />
@@ -119,7 +119,7 @@ export default function EditEventForm(props) {
                         helperText={touched.elocation ? errors.elocation : ""}
                         error={touched.elocation && Boolean(errors.elocation)}
                         value={elocation}
-                        onChange={change.bind(null, "elocation")}
+                        onChange={change.bind(null, "location")}
                     />
                     <TextField
                         id="imageUrl"
@@ -138,5 +138,3 @@ export default function EditEventForm(props) {
         </ThemeProvider>
     )
 }
-
-
