+++ src/components/Forms/RegisterEventForm.js
@@ -2,7 +2,7 @@ import React from "react"
 import { Button, TextField } from "@material-ui/core";
 import Grid from '@material-ui/core/Grid';
 import Typography from '@material-ui/core/Typography';
-import RadioGroupButtons from '../../RadioGroupButtons';
+import RadioGroupButtons from '../RadioGroupButtons';
 
 export default function RegisterEventForm(props) {
     const {
